backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (MongoDB)
motor==3.3.2
pymongo[srv]==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Configuration
pydantic-settings==2.2.1

# HTTP Client (for testing / webhooks)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
